Move assertIntervals and shorten the string length
